LEPL-23: промежуточные изменения по таске. Поправил AdminCourseMapperTest.
